requirements.txt
Django==2.1.1
django-auto-one-to-one==3.1.0
django-extensions==2.1.2
djangorestframework==3.8.2
PyJWT==1.6.4
python-dateutil==2.7.3
pytz==2018.5
six==1.11.0
